ADDED: PdfIdentityEncoding which does not do anything useful FIXED: Encodings have now an autodelete property, which is convinient for parameterized Identity encodings FIXED: A few encoding bugs. FIXED: A few compiler warnings in PdfTable and PdfPainter.
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@772 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfTable.cpp
@@ -114,8 +114,10 @@ void PdfTable::Draw( double dX, double dY, PdfPainter* pPainter, const PdfRect &
                               pdColWidths, pdRowHeights,
                               &dWidth, &dHeight );
  
-    if( !(!rClipRect.GetBottom() && !rClipRect.GetLeft() &&
-          !rClipRect.GetWidth() && !rClipRect.GetWidth()) ) 
+    if( !(!static_cast<int>(rClipRect.GetBottom()) && 
+          !static_cast<int>(rClipRect.GetLeft()) &&
+          !static_cast<int>(rClipRect.GetWidth()) && 
+          !static_cast<int>(rClipRect.GetWidth())) ) 
         m_curClipRect = rClipRect;
     else
     {
